#Conflicts:
#	internal/core/window.rs

+++ api/node/rust/types/brush.rs
@@ -159,17 +159,11 @@ impl SlintRgbaColor {
     /// Returns the color as string in hex representation e.g. `#000000` for black.
     #[napi]
     pub fn to_string(&self) -> String {
-        let alpha = self.alpha() as u8;
+        let alpha = self.alpha();
         if alpha == 255 {
-            format!("#{:02x}{:02x}{:02x}", self.red() as u8, self.green() as u8, self.blue() as u8)
+            format!("#{:02x}{:02x}{:02x}", self.red(), self.green(), self.blue())
         } else {
-            format!(
-                "#{:02x}{:02x}{:02x}{:02x}",
-                self.red() as u8,
-                self.green() as u8,
-                self.blue() as u8,
-                alpha
-            )
+            format!("#{:02x}{:02x}{:02x}{:02x}", self.red(), self.green(), self.blue(), alpha)
         }
     }
 }
